+++ Sources/StreamChat/APIClient/Endpoints/Payloads/MessagePayloads.swift
@@ -39,6 +39,7 @@ enum MessagePayloadsCodingKeys: String, CodingKey, CaseIterable {
     case mml
     case imageLabels = "image_labels"
     case shadowed
+    case moderationDetails = "moderation_details"
 }
 
 extension MessagePayload {
@@ -85,6 +86,7 @@ class MessagePayload: Decodable {
     let isShadowed: Bool
     let translations: [TranslationLanguage: String]?
     let originalLanguage: String?
+    let moderationDetails: MessageModerationDetailsPayload?
 
     var pinned: Bool
     var pinnedBy: UserPayload?
@@ -149,6 +151,7 @@ class MessagePayload: Decodable {
         let i18n = try container.decodeIfPresent(MessageTranslationsPayload.self, forKey: .i18n)
         translations = i18n?.translated
         originalLanguage = i18n?.originalLanguage
+        moderationDetails = try container.decodeIfPresent(MessageModerationDetailsPayload.self, forKey: .moderationDetails)
     }
 
     init(
@@ -183,7 +186,8 @@ class MessagePayload: Decodable {
         pinnedAt: Date? = nil,
         pinExpires: Date? = nil,
         translations: [TranslationLanguage: String]? = nil,
-        originalLanguage: String? = nil
+        originalLanguage: String? = nil,
+        moderationDetails: MessageModerationDetailsPayload? = nil
     ) {
         self.id = id
         self.cid = cid
@@ -217,6 +221,7 @@ class MessagePayload: Decodable {
         self.quotedMessageId = quotedMessageId
         self.translations = translations
         self.originalLanguage = originalLanguage
+        self.moderationDetails = moderationDetails
     }
 }
 
@@ -307,8 +312,6 @@ struct MessageListPayload: Decodable {
 struct MessageReactionsPayload: Decodable {
     let reactions: [MessageReactionPayload]
 }
-
-// TODO: Command???
 
 /// A command in a message, e.g. /giphy.
 public struct Command: Codable, Hashable {