+++ TestTools/StreamChatTestTools/Mocks/StreamChat/Repositories/MessageRepository_Mock.swift
@@ -59,7 +59,11 @@ final class MessageRepository_Mock: MessageRepository, Spy {
         completion?(saveSuccessfullyDeletedMessageError)
     }
 
-    override func updateMessage(withID id: MessageId, localState: LocalMessageState?, isBounced: Bool? = nil, completion: @escaping () -> Void) {
+    override func updateMessage(
+        withID id: MessageId,
+        localState: LocalMessageState?,
+        completion: @escaping () -> Void
+    ) {
         record()
         updatedMessageLocalState = localState
         completion()