+++ Tests/StreamChatTests/Workers/Background/MessageEditor_Tests.swift
@@ -78,7 +78,7 @@ final class MessageEditor_Tests: XCTestCase {
             }))
         }
 
-        XCTAssertCall("updateMessage(withID:localState:isBounced:completion:)", on: messageRepository, times: 1)
+        XCTAssertCall("updateMessage(withID:localState:completion:)", on: messageRepository, times: 1)
     }
 
     func test_editorSyncsMessage_whenMessageChangesToPendingSyncAndHasAttachmentsUploadedFromServer() throws {
@@ -114,7 +114,7 @@ final class MessageEditor_Tests: XCTestCase {
                 $0.endpoint == AnyEndpoint(.editMessage(payload: message1Payload))
             }.count, 1)
         }
-        XCTAssertCall("updateMessage(withID:localState:isBounced:completion:)", on: messageRepository, times: 1)
+        XCTAssertCall("updateMessage(withID:localState:completion:)", on: messageRepository, times: 1)
     }
 
     func test_editorSyncsMessage_whenAttachmentFinishesUploading() throws {
@@ -154,7 +154,7 @@ final class MessageEditor_Tests: XCTestCase {
 
         // Then
         XCTAssertTrue(apiClient.request_allRecordedCalls.count == 1)
-        XCTAssertCall("updateMessage(withID:localState:isBounced:completion:)", on: messageRepository, times: 1)
+        XCTAssertCall("updateMessage(withID:localState:completion:)", on: messageRepository, times: 1)
     }
 
     func test_editor_changesMessageStates_whenSyncingSucceeds() throws {
@@ -169,7 +169,7 @@ final class MessageEditor_Tests: XCTestCase {
 
         // Check the state is eventually changed to `syncing`
         AssertAsync.willBeEqual(messageRepository.updatedMessageLocalState, .syncing)
-        XCTAssertCall("updateMessage(withID:localState:isBounced:completion:)", on: messageRepository, times: 1)
+        XCTAssertCall("updateMessage(withID:localState:completion:)", on: messageRepository, times: 1)
 
         // Wait for the API call to be initiated
         AssertAsync.willBeTrue(apiClient.request_endpoint != nil)
@@ -180,7 +180,7 @@ final class MessageEditor_Tests: XCTestCase {
 
         // Check the state is eventually changed to `nil`
         AssertAsync.willBeEqual(messageRepository.updatedMessageLocalState, nil)
-        XCTAssertCall("updateMessage(withID:localState:isBounced:completion:)", on: messageRepository, times: 2)
+        XCTAssertCall("updateMessage(withID:localState:completion:)", on: messageRepository, times: 2)
     }
 
     func test_editor_changesMessageStates_whenSyncingFails() throws {
@@ -195,7 +195,7 @@ final class MessageEditor_Tests: XCTestCase {
 
         // Check the state is eventually changed to `syncing`
         AssertAsync.willBeEqual(messageRepository.updatedMessageLocalState, .syncing)
-        XCTAssertCall("updateMessage(withID:localState:isBounced:completion:)", on: messageRepository, times: 1)
+        XCTAssertCall("updateMessage(withID:localState:completion:)", on: messageRepository, times: 1)
 
         // Wait for the API call to be initiated
         AssertAsync.willBeTrue(apiClient.request_endpoint != nil)
@@ -205,7 +205,7 @@ final class MessageEditor_Tests: XCTestCase {
 
         // Check the state is eventually changed to `syncingFailed`
         AssertAsync.willBeEqual(messageRepository.updatedMessageLocalState, .syncingFailed)
-        XCTAssertCall("updateMessage(withID:localState:isBounced:completion:)", on: messageRepository, times: 2)
+        XCTAssertCall("updateMessage(withID:localState:completion:)", on: messageRepository, times: 2)
     }
 
     func test_editor_doesNotRetainItself() throws {
@@ -225,7 +225,7 @@ final class MessageEditor_Tests: XCTestCase {
             Assert.willBeTrue(self.apiClient.request_endpoint != nil)
         }
 
-        XCTAssertCall("updateMessage(withID:localState:isBounced:completion:)", on: messageRepository, times: 1)
+        XCTAssertCall("updateMessage(withID:localState:completion:)", on: messageRepository, times: 1)
         // Assert editor can be released even though response hasn't come yet
         AssertAsync.canBeReleased(&editor)
     }